Fix armv7 build: use usize atomics for stats counters

+++ src/server.zig
@@ -72,10 +72,10 @@ const Pool = struct {
 /// Atomic operational counters. Updated off the per-connection hot path with
 /// relaxed (monotonic) ordering — they are observability hints, not a barrier.
 const Stats = struct {
-    accepted: std.atomic.Value(u64) = .init(0),
-    refused: std.atomic.Value(u64) = .init(0),
-    active: std.atomic.Value(u64) = .init(0),
-    connect_failures: std.atomic.Value(u64) = .init(0),
+    accepted: std.atomic.Value(usize) = .init(0),
+    refused: std.atomic.Value(usize) = .init(0),
+    active: std.atomic.Value(usize) = .init(0),
+    connect_failures: std.atomic.Value(usize) = .init(0),
 };
 
 pub const Server = struct {
